refactor(l5): move distill, embedding-repair and graph-derivation handlers from src/ into server/maintenance (issue 864)
Moved paths (old -> new):

- src/distill-handler.ts -> server/maintenance/distill-handler.ts
- src/embedding-repair-handler.ts -> server/maintenance/embedding-repair-handler.ts
- src/graph-derivation-handler.ts -> server/maintenance/graph-derivation-handler.ts

distill-handler.ts was 602 lines and its sweep carried three responsibilities
that each hold their own state and error path, so it is split into siblings
named for what they do:

- server/maintenance/distill-claim.ts -- claim one batch of undistilled turns,
  traced when a recorder is present.
- server/maintenance/distill-persist.ts -- the embedding batch and the
  transactional write (insert candidates, stamp their turns).
- server/maintenance/distill-backlog.ts -- classify an empty claim as a drained
  queue or as the pathological case where work remained.

graph-derivation-handler.ts keeps its terminal validation gates in one new
module-scope function, validateDerivationJob, instead of inline in the handler
closure.

No exported function signature changed for any caller outside the moved set;
runDistillSweep, makeMemoryDistillHandler, buildMemoryDistillEnqueue,
createEmbeddingRepairHandler, buildEmbeddingRepairHandlers,
makeGraphDerivationHandler, selectSourcesNeedingDerivation,
buildGraphDerivationEnqueue and enqueueGraphDerivationJobs are unchanged.

Each old src/ path is now an L5 shim per M2, so src/maintenance-bootstrap.ts,
src/maintenance-sweep.ts, src/distill-exchange-run.ts, scripts/distill-run.ts
and scripts/dream-cycle.ts keep their import lines untouched.

Imports of modules already moved point at the server/ path directly
(background-tracing, distill-window, distiller, embedding-targets); imports of
modules still in src/ are relative per M3 (../../src/<x>.ts).

Runtime closure of src/ under server/main.ts: 48 -> 43.
